// albiononline/src/types.ts
export interface Character {
  name: string;
  capital: number;
}